separate plotting and data prep and rename file
